requirements.txt
Flask
marshmallow
simplejson
flasgger
apispec
PyYAML
git+https://git@github.com/wobeng/aws-helper.git@master
git+https://git@github.com/wobeng/py-helper.git@master